+++ highlighting/lib/highlighting.dart
@@ -1,8 +1,8 @@
-import 'src/highlightv2.dart';
+import 'src/highlight.dart';
 
-export 'package:highlighting/src/highlightv2.dart';
+export 'package:highlighting/src/highlight.dart';
 export 'package:highlighting/src/mode.dart';
 export 'package:highlighting/src/node.dart';
 export 'package:highlighting/src/result.dart';
 
-final highlight = HighlightV2();
+final highlight = Highlight();