+++ highlighting/lib/languages/django.dart
@@ -1,6 +1,11 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-import '../src/language_definition_parts.dart';
+// ignore_for_file: lines_longer_than_80_chars
+// ignore_for_file: prefer_single_quotes
+// ignore_for_file: unnecessary_raw_strings
+// ignore_for_file: use_raw_strings
+
+import '../src/language_definition_common.dart';
 
 final django = Language(
   id: "django",