+++ highlighting/lib/src/common_modes.g.dart
@@ -1,5 +1,12 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
+// ignore_for_file: lines_longer_than_80_chars
+// ignore_for_file: prefer_single_quotes
+// ignore_for_file: unnecessary_raw_strings
+// ignore_for_file: use_raw_strings
+
+// ignore_for_file: non_constant_identifier_names
+
 import 'mode.dart';
 
 final BACKSLASH_ESCAPE = Mode(