+++ highlighting/lib/src/language_definition_common.dart
@@ -1,7 +1,6 @@
 export 'callbacks.dart';
-export 'common_modes.dart';
+export 'common_modes.g.dart';
 export 'language.dart';
+export 'languages/mathematica_symbols.g.dart';
 export 'mode.dart';
 export 'nulls.dart';
-
-export 'languages/mathematica_symbols.dart';