+++ tool/src/languages/mathematica.ts
@@ -1,6 +1,6 @@
 import { readFileSync, writeFileSync } from "fs";
 
-import { NOTICE_COMMENT } from "../common.js";
+import { IGNORE_FOR_FILE_COMMENT, NOTICE_COMMENT } from "../common.js";
 
 export function portMathematicaSpecific(): void {
   const fileName = "../node_modules/highlight.js/lib/languages/mathematica.js";
@@ -13,12 +13,13 @@ export function portMathematicaSpecific(): void {
 
   const output =
     NOTICE_COMMENT +
+    IGNORE_FOR_FILE_COMMENT +
     "final SYSTEM_SYMBOLS = " +
     symbolSet[1].replace(/\[/, "{").replace(/\]/, "}").replace(/\s"/g, 'r"') +
     ";";
 
   writeFileSync(
-    "../../highlighting/lib/src/languages/mathematica_symbols.dart",
+    "../../highlighting/lib/src/languages/mathematica_symbols.g.dart",
     output,
   );
 }